Move converters to their own package
